pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "django-blackbox"
version = "0.1.0"
description = "A Django app that captures and tracks 5xx server errors with rich metadata and traceable incident IDs"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [
    { name = "Django Black Box Contributors" }
]
keywords = ["django", "error-tracking", "monitoring", "logging", "incidents", "blackbox"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Framework :: Django",
    "Framework :: Django :: 3.2",
    "Framework :: Django :: 4.0",
    "Framework :: Django :: 4.1",
    "Framework :: Django :: 4.2",
    "Framework :: Django :: 5.0",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Internet :: WWW/HTTP :: Dynamic Content",
    "Topic :: System :: Logging",
]

dependencies = [
    "Django>=3.2",
    "djangorestframework>=3.14",
]

[project.optional-dependencies]
ulid = ["ulid-py>=1.1.0"]
dev = [
    "pytest>=7.0",
    "pytest-django>=4.5.0",
    "pytest-cov>=4.0.0",
    "coverage>=6.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["django_blackbox*"]
exclude = ["tests*"]

[tool.setuptools.package-data]
django_blackbox = ["py.typed"]

